Exceptions were added

+++ SprintFinal/Product/src/main/java/com/finalSprint/product/service/ProductServices.java
@@ -6,9 +6,10 @@ import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import com.finalSprint.product.exceptions.InvalidValueException;
 import com.finalSprint.product.model.Product;
 import com.finalSprint.product.repository.ProductRepository;
- 
+
 @Service
 public class ProductServices {
 
@@ -21,13 +22,18 @@ public class ProductServices {
 	}
 
 	// filter by id
-	public Product getProductById(long pid) {
-		return repository.findById(pid).get();
+	public Product getProductById(long pid) throws InvalidValueException, Exception {
+		return repository.findById(pid).orElseThrow(() -> new InvalidValueException("Product not found"));
 	}
 
 	// filter by name
-	public Optional<Product> getProductByName(String name) {
-		return repository.findByProductName(name);
+	public Optional<Product> getProductByName(String name) throws InvalidValueException, Exception {
+		Optional<Product> product = repository.findByProductName(name);
+		if (product.isPresent()) {
+			return product;
+		} else {
+			throw new InvalidValueException("Product not found");
+		}
 	}
 
 	// filter by price
@@ -39,9 +45,14 @@ public class ProductServices {
 	public List<Product> getProductLessThan(double price) {
 		return repository.findByProductPriceLessThan(price);
 	}
-	
+
 	// update product
-	public Product UpdateProduct(long pid, Product productDetails) {
+	public Product UpdateProduct(long pid, Product productDetails) throws InvalidValueException, Exception {
+		if (productDetails.getProductName() == null | productDetails.getProductPrice() == 0
+				| productDetails.getQuantity() == 0) {
+			throw new NullPointerException("Product cannot be null");
+		}
+
 		Product product = repository.findById(pid).get();
 
 		product.setProductName(productDetails.getProductName());
@@ -52,14 +63,25 @@ public class ProductServices {
 	}
 
 	// delete product
-	public void deleteById(long pid) {
+	public void deleteById(long pid) throws InvalidValueException, Exception {
+		if (!repository.existsById(pid)) {
+			throw new InvalidValueException("Product not found");
+		}
 		repository.deleteById(pid);
 	}
 
 	// add products
-	public Product addOrUpdateProduct(Product product) {
+	public Product addOrUpdateProduct(Product product) throws InvalidValueException, Exception {
+
+		if (product.getProductName() == null | product.getProductPrice() == 0 | product.getQuantity() == 0) {
+			throw new NullPointerException("Product cannot be null");
+		}
+
 		Optional<Product> existingProduct = repository.findByProductName(product.getProductName());
 
+		if (product.getQuantity() < 0) {
+			throw new InvalidValueException("Quantity cannot be negative.");
+		}
 		if (existingProduct.isPresent()) {
 			Product exprod = existingProduct.get();
 			exprod.setQuantity(exprod.getQuantity() + product.getQuantity());
@@ -67,10 +89,16 @@ public class ProductServices {
 		} else {
 			return repository.save(product);
 		}
+
 	}
 
 	// subtract product
-	public Optional<Product> subtractProduct(Product product) {
+	public Optional<Product> subtractProduct(Product product) throws InvalidValueException, Exception {
+
+		if (product.getProductName() == null | product.getProductPrice() == 0 | product.getQuantity() == 0) {
+			throw new NullPointerException("Product cannot be null");
+		}
+
 		Optional<Product> existingProduct = repository.findByProductName(product.getProductName());
 
 		if (existingProduct.isPresent()) {
@@ -80,15 +108,13 @@ public class ProductServices {
 				return Optional.of(repository.save(exproduct));
 			} else {
 				// when quantity to subtract is greater than existing quantity
-				System.out.println("Insufficient quantity to subtract.");
-				return Optional.empty();
+				throw new InvalidValueException("Insufficient quantity");
+
 			}
 		} else {
 			// when product is not found
-			System.out.println("Product not found.");
-			return Optional.empty();
+			throw new InvalidValueException("Product not found");
 		}
 	}
-
 
 }
